+++ packages/core/src/components/ChoiceList/Choice.scss
@@ -1,6 +1,21 @@
 @import '../../settings/index';
 
 /*
+Fieldset
+
+Markup:
+<fieldset class="ds-c-fieldset">...</fieldset>
+*/
+.ds-c-fieldset {
+  border: 0;
+  // We apply the same margin as a label has, since in the context of a fieldset
+  // this would be a <legend>. See .ds-c-fieldset > .ds-c-label for more info.
+  margin: $spacer-3 0 0;
+  min-width: 0;
+  padding: 0;
+}
+
+/*
 Checkbox & Radio
 
 @status prototype
