+++ packages/core/src/components/_index.scss
@@ -15,8 +15,8 @@ Styleguide components
 @import 'Button/Button';
 @import 'ChoiceList/Choice';
 @import 'ChoiceList/Select';
-@import 'Form/FormField';
-@import 'Form/FormLabel';
+@import 'FormLabel/FormLabel';
 @import 'List/List';
 @import 'Table/Table';
+@import 'TextField/TextField';
 @import 'VerticalNav/VerticalNav';
